* 'develop' of https://gitlab.com/jusfoun-bigdata/hookah:
  站内信
  站内信
  实名认证 zc
  zc
  优化
  供应商管理完成
  zc
  bug
  邮件下发地址更换及代码完善
  fix
  fix bug
  结算处理
  下拉样式
  邮件下发地址更换及代码完善
  测试数据

+++ hookah-pay/src/main/java/com.jusfoun.hookah.pay/service/impl/MgMoneyInOutLogServiceImpl.java
@@ -40,7 +40,7 @@ public class MgMoneyInOutLogServiceImpl extends GenericMongoServiceImpl<MgMoneyI
         }
         mgMoneyInOutLog.setPayAccountRecordId(id);
         mgMoneyInOutLog.setUserId(moneyInOutBo.getUserId());
-        mgMoneyInOutLog.setTransferType(moneyInOutBo.getOperatorType());
+        mgMoneyInOutLog.setTransferType((byte)moneyInOutBo.getOperatorType().intValue());
         mgMoneyInOutLog.setAddTime(new Date());
         mgMoneyInOutLog.setSendArgs(sendArgs);
         mongoTemplate.insert(mgMoneyInOutLog);
